#hoale Product_ProductGroup

+++ src/app/pages/subscription/subscription.component.ts
@@ -5,8 +5,9 @@ import { LoadSubscriptionAction, DeleteSubscriptionAction, AddSubscriptionAction
 import { Observable } from 'rxjs';
 import { Store } from '@ngrx/store';
 import { from } from 'rxjs';
-import { HttpErrorResponse, HttpHeaders, HttpClient } from '@angular/common/http';
+import { HttpHeaders, HttpClient } from '@angular/common/http';
 import { ServerHttpService } from 'src/app/Services/server-http.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-subscription',
@@ -29,6 +30,8 @@ export class SubscriptionComponent implements OnInit {
   subscriptionEdit : Subscription = {company:{id:0,address:'',logo:'',phone_number:'',store_name:''}, 
   company_id:0,plan_id:'', trial_period_days:'',plan:{id:0,name:'',price:0}};
 
+  err : boolean = false;
+
   constructor(private store: Store<AppState>,
     private http : HttpClient,
     private httpServer: ServerHttpService) { }
@@ -48,20 +51,38 @@ export class SubscriptionComponent implements OnInit {
     this.subscriptionItems.subscribe(data => console.log(data));
     this.http.get<any>('https://seekproduct-api.misavu.net/api/user/company/my-company/',{
       headers: new HttpHeaders({
-      Authorization: 'JWT ' + this.httpServer.tokenLogin,
+      Authorization: 'JWT ' + localStorage.getItem('TOKEN'),
     })}).subscribe(data => {
       this.company = data.results;
+      this.company_id = data.results[0].id;
+      this.selectCompany = data.results[0].id;
+      console.log(data.results[0].id);
   });
 
   this.http.get<any>('https://seekproduct-api.misavu.net/api/plan/').subscribe(data => {
       this.planArr = data;
+      this.selectPlan = data[0].id;
+      this.plan_id = data[0].id;
+      console.log(data[0].id);
   });
-  
   }
 
   deleteSubscriptionItem(id : any){
     console.log("company id : " + id);
-    this.store.dispatch(new DeleteSubscriptionAction(id));
+
+    Swal.fire({
+      title: 'Are you sure?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.store.dispatch(new DeleteSubscriptionAction(id));
+      }
+    });
+    
   }
   openDialog(){
     this.show = !this.show;
@@ -69,9 +90,32 @@ export class SubscriptionComponent implements OnInit {
   changeData (event : any){
     this[event.target.id] = event.target.value;
     console.log(event.target.id + ": " + this[event.target.id]);
-    if(event.target.id === 'company_id'){
+    if(event.target.id == 'company_id'){
+      this.subscriptionItems.subscribe(data => {
+
+        for(var i = 0; i < data.length ; i++){
+          if(data[i].company.id == event.target.value){
+            console.log(data[i].company.id);
+            this.err = true;
+            break;
+          }
+          else {
+            this.err = false;
+          }
+        }
+        // data.forEach((value : any,key)=> {
+        //   if(value.company.id == event.target.value){
+        //     this.err = true;
+        //   }
+        // });
+      });
+    }
+  }
+  submit(){
+    console.log(this.item);
+    if(this.checkEdit === false){
       this.company.forEach((value,key)=>{
-        if(value.id == event.target.value){
+        if(value.id == this.company_id){
           console.log(value);
           this.item.company.id = value.id;
           this.item.company.address = value.address;
@@ -80,51 +124,59 @@ export class SubscriptionComponent implements OnInit {
           this.item.company.phone_number = value.phone_number;
         } 
       });
-    }
-    if(event.target.id === 'plan_id'){
       this.planArr.forEach((value,key)=>{
-        if(value.id == event.target.value){
+        if(value.id == this.plan_id){
           console.log(value);
           this.item.plan.id = value.id;
           this.item.plan.name = value.name;
           this.item.plan.price = value.price;
         }
       });
-    }
-  }
-  submit(){
-    console.log(this.item);
-    if(this.checkEdit === false){
-    this.item.company_id = parseInt(this.company_id);
-    this.item.plan_id = this.plan_id;
-    this.item.trial_period_days = this.trial_period_days;
-    
-    this.store.dispatch(new AddSubscriptionAction(this.item));
-    this.item = {company:{id:0,address:'',logo:'',phone_number:'',store_name:''}, 
-    company_id:0,plan_id:'', trial_period_days:'',plan:{id:0,name:'',price:0}};
-    this.company_id = '';
-    this.plan_id = '';
-    this.trial_period_days = '';
-    }
-    if(this.checkEdit === true){
-      // this.subscriptionEdit.plan_id = this.plan_id;
-      // this.subscriptionEdit.trial_period_days = this.trial_period_days;
-      // this.subscriptionEdit.company_id = parseInt(this.company_id);
       this.item.company_id = parseInt(this.company_id);
       this.item.plan_id = this.plan_id;
       this.item.trial_period_days = this.trial_period_days;
-      this.planArr.forEach((value2,key)=>{
-        if(value2.id == this.plan_id){
-          this.item.plan = value2;
-        }
-      });      
-      this.store.dispatch(new UpdateSubscriptionAction(this.item));
-      this.checkEdit = false;
-    }
-    this.openDialog();
+      this.store.dispatch(new AddSubscriptionAction(this.item));
+      this.item = {company:{id:0,address:'',logo:'',phone_number:'',store_name:''}, 
+        company_id:0,plan_id:'', trial_period_days:'',plan:{id:0,name:'',price:0}};
+    } else if(this.checkEdit === true){
+              this.item.company_id = parseInt(this.company_id);
+              this.item.plan_id = this.plan_id;
+              this.item.trial_period_days = this.trial_period_days;
+              this.planArr.forEach((value2,key)=>{
+                if(value2.id == this.plan_id){
+                  this.item.plan = value2;
+                }
+              });
+              this.company.forEach((value,key)=>{
+                if(value.id == this.company_id){
+                  this.item.company = value;
+                }
+              });
+              Swal.fire({
+                title: 'Are you sure?',
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, Update it!'
+              }).then((result) => {
+                if (result.value) {
+                  this.store.dispatch(new UpdateSubscriptionAction(this.item));
+                  this.item = {company:{id:0,address:'',logo:'',phone_number:'',store_name:''}, 
+                        company_id:0,plan_id:'', trial_period_days:'',plan:{id:0,name:'',price:0}};
+                }
+              });     
+              this.checkEdit = false;
+            }
+             
+            this.company_id = '';
+            this.plan_id = '';
+            this.trial_period_days = '';
+            this.openDialog();
   }
 
   selectPlan : number;
+  selectCompany : number;
   EditSubscriptionItem(id : any){
     console.log(id);
     this.checkEdit = true;
@@ -132,26 +184,16 @@ export class SubscriptionComponent implements OnInit {
       data.forEach((value : any,key)=> {
         if(value.id == id){
           this.subscriptionEdit = value;
-          console.log(value.plan.price);
-          this.company.forEach((value1,key) => {
-            if(value1.id == value.company.id){
-              this.company_id = value1.store_name;
-              console.log(value1.store_name);
-            }
-          });
-
-          this.planArr.forEach((value2,key)=>{
-            if(value2.id == value.plan.id){
-              console.log(value2.name);
-              
-            }
-          });
         }
       });
     });
     this.selectPlan = this.subscriptionEdit.plan.id;
+    this.selectCompany =this.subscriptionEdit.company.id;
     this.company_id =this.subscriptionEdit.company.id + '';
-    this.openDialog();
+
+    if(this.show === false){
+      this.openDialog();
+    }
     console.log(this.subscriptionEdit);
   }
 }
